Move Threesizer to own dir
